requirements.txt
requests==2.32.5
playwright==1.58.0
openai==2.24.0
python-dotenv==1.2.2
PySide6==6.10.2
